refactor: ui as a module
