+++ src/index.js
@@ -1,7 +1,7 @@
-import Store from "./services/store.js";
-import UI from "./controllers/ui.js";
-import VirtualKeyboard from "./components/virtual-keyboard.js";
 import Wordless from "./controllers/wordless.js";
+import GA from "./modules/analytics.js";
+import UI from "./modules/ui.js";
+import Store from "./services/store.js";
 
 window.wordless = {
   store: Store,
@@ -14,10 +14,8 @@ window.wordless = {
 };
 
 document.addEventListener("DOMContentLoaded", () => {
-  // register virtual keyboard component
-  customElements.define("virtual-keyboard", VirtualKeyboard);
-  // initialize UI modals
-  UI.initModals();
+  // load UI components
+  UI.init(GA);
   // start wordless
   wordless.start();
 });
